minor fixup for camlp5 8.00
there are now PPX attributes in many places, so (for example) what was
a pair (pattern * expression) is now a triple (pattern * expression *
attributes).  Typically the attributes are empty in generated code.
Since this is a spot that can have a meta-variable in quotations, we
have to wrap the empty-list in a "vala".  But again, this is pretty
much a boilerplate value.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch camlp5-8.00-patches
# Changes to be committed:
#	modified:   pa_ulex.ml
#
# Untracked files:
#	_build/
#