Port to OCaml 5.0.0, bump version to 1.3.
* minor fixup for camlp5 8.00

there are now PPX attributes in many places, so (for example) what was
a pair (pattern * expression) is now a triple (pattern * expression *
attributes).  Typically the attributes are empty in generated code.
Since this is a spot that can have a meta-variable in quotations, we
have to wrap the empty-list in a "vala".  But again, this is pretty
much a boilerplate value.

Co-authored-by: Chet Murthy <chetsky@gmail.com>